Created variable mapping structure
