+++ crates/sym/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod aiger;
 mod buf;
 mod convert;
+mod eval;
 mod pcode;
 mod sym;
 
@@ -8,6 +9,7 @@ pub use crate::buf::*;
 pub use crate::convert::concretize;
 pub use crate::convert::concretize_into;
 pub use crate::convert::ConcreteValue;
+pub use crate::eval::*;
 pub use crate::sym::*;
 
 #[cfg(test)]
