+++ crates/sym/src/tests.rs
@@ -3,4 +3,5 @@ mod bit;
 mod bitbuf;
 mod bitvec;
 mod convert;
+mod eval;
 mod pcode;
